Refac: 도메인별로 ErrorCode 분리
# Conflicts:
#	src/main/java/org/sopt/global/common/ErrorCode.java
#	src/main/java/org/sopt/service/PostService.java
#	src/main/java/org/sopt/service/UserService.java
